CR requests Rename class and move interfaces to folder
