--- RepoDb.Core/RepoDb/IIdOptions.cs
@@ -1,7 +1,0 @@
-namespace RepoDb
-{
-    public interface IIdOptions<T> where T : class
-    {
-        IIdOptions<T> Column(string column);
-    }
-}
