+++ RepoDb.Core/RepoDb/IdentityOptions.cs
@@ -1,18 +1,19 @@
 using System;
 using System.Linq.Expressions;
+using RepoDb.Interfaces;
 
 namespace RepoDb
 {
-    internal class IdOptions<T> : IIdOptions<T> where T : class
+    internal class IdentityOptions<T> : IIdentityOptions<T> where T : class
     {
         private readonly Expression<Func<T, object>> m_expression;
 
-        public IdOptions(Expression<Func<T, object>> expression)
+        public IdentityOptions(Expression<Func<T, object>> expression)
         {
             m_expression = expression;
         }
 
-        public IIdOptions<T> Column(string column)
+        public IIdentityOptions<T> Column(string column)
         {
             PropertyMapper.Add(m_expression, column);
             return this;
