+++ RepoDb.Core/RepoDb/Interfaces/IIdentityOptions.cs
@@ -1,0 +1,7 @@
+namespace RepoDb.Interfaces
+{
+    public interface IIdentityOptions<T> where T : class
+    {
+        IIdentityOptions<T> Column(string column);
+    }
+}
